+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Fork Modification
+
+* Autoplay configuration
+
+
 # Changelog
 
 * now requires PHP 5.4.0+